Fixed warnings

+++ src/sound/sound_server.cpp
@@ -287,7 +287,7 @@ local int MixSampleToStereo32(Sample* sample,int index,unsigned char volume,
 {
     int ri;				// read index
     int wi;				// write index
-    int length;
+    unsigned length;
     int v;
     int local_volume;
     unsigned char left;
@@ -738,7 +738,7 @@ local int MixChannelsToStereo32(int* buffer,int size)
 		    buffer,size);
 	    Channels[channel].Point+=i;
 
-	    if( Channels[channel].Point>=Channels[channel].Sample->Length ) {
+	    if( (unsigned)Channels[channel].Point>=Channels[channel].Sample->Length ) {
 		// free channel as soon as possible (before playing)
 		// useful in multithreading
 		DebugLevel3("End playing request from %p\n" _C_
